add localStorage and divide form into partials
