+++ src/js/card-object.js
@@ -1,4 +1,4 @@
-const cardData = {
+let cardData = {
     palette: "1",
     name: "",
     job: "",
