+++ src/js/create-card.js
@@ -1,11 +1,4 @@
-const btnCreate = document.querySelector ('.js_btnCreate');
-const sectionCreated = document.querySelector('.js-section-created');
-
-
-
-function handleCreate (event){
-    event.preventDefault();
-    
+const postData = () =>{
     fetch('https://dev.adalab.es/api/card/',{
         method:'POST',
         headers:{'content-type':'application/json'},
@@ -19,10 +12,10 @@ function handleCreate (event){
             btnCreate.classList.remove('btn-create');
             btnCreate.classList.add('btn-create_grey');
             sectionCreated.classList.remove('collapsed');
-            localStorage.setItem('newCard', JSON.stringify(cardData));
             cardLink.innerHTML = dataResponse.cardURL;
             cardLink.href = dataResponse.cardURL;
             createError.innerHTML = '';
+            
         }else if(dataResponse.error.toLowerCase().includes('mandatory field')){
             createError.innerHTML = 'Es necesario rellenar todos los campos.';          
         }else if(dataResponse.error.includes('ER_DATA_TOO_LONG')){
@@ -30,9 +23,12 @@ function handleCreate (event){
         }
         
     })
-    // .catch((error)=> console.log(`Ha sucedido un error:${error}`));
+}
 
+
+function handleCreate (event){
+    event.preventDefault();
+    postData();
 }
 
 btnCreate.addEventListener('click', handleCreate);
-
