+++ src/js/main.js
@@ -2,9 +2,11 @@
 
 import "./variables.js";
 import "./card-object.js";
-import "./form.js";
+import "./handle-data.js";
 import "./get-avatar.js";
 import "./collapsable.js";
+import "./reset.js"
 import "./create-card.js";
 import "./share.js";
+import "./init.js";
