+++ src/partials/design.html
@@ -12,19 +12,19 @@
         <label for="palette">COLORES</label>
         <section class="fieldset__design--colors">
             <div class="color-one">
-                <input type="radio" value="1" name="palette" class="onecol js-onecol" checked>
+                <input type="radio" value="1" name="palette" class="onecol js-option" checked>
                 <div class="color-one__1"></div>
                 <div class="color-one__2"></div>
                 <div class="color-one__3"></div>
             </div>
             <div class="color-two">
-                <input type="radio" value="2" name="palette" class="onecol">
+                <input type="radio" value="2" name="palette" class="onecol js-option">
                 <div class="color-two__1"></div>
                 <div class="color-two__2"></div>
                 <div class="color-two__3"></div>
             </div>
             <div class="color-three">
-                <input type="radio" value="3" name="palette" class="onecol">
+                <input type="radio" value="3" name="palette" class="onecol js-option">
                 <div class="color-three__1"></div>
                 <div class="color-three__2"></div>
                 <div class="color-three__3"></div>
